requirements.txt
# Required dependencies
beautifulsoup4
requests
python-dotenv
pandas
scikit-learn

# Optional: Only needed if use_ai=True
# anthropic
